Check unzipped directory doesn't already exist to avoid error

+++ lib/src/sample_data.dart
@@ -28,7 +28,7 @@ Future<void> downloadSampleData(List<String> portalItemIds) async {
 
   for (final itemId in portalItemIds) {
     // create a portal item to ensure it exists and load to access properties
-    var portalItem =
+    final portalItem =
         PortalItem.withUri(Uri.parse('$portal/home/item.html?id=$itemId'));
     if (portalItem != null) {
       await portalItem.load();
@@ -42,10 +42,14 @@ Future<void> downloadSampleData(List<String> portalItemIds) async {
         // if the data is a zip we need to extract it
         // save all files to the device app directory in a directory with the item name without the zip extension
         final nameWithoutExt = itemName.replaceFirst(RegExp(r'.zip$'), '');
-        final dir = Directory.fromUri(Uri.parse('$appDirPath/$nameWithoutExt'));
-        await ZipFile.extractToDirectory(zipFile: file, destinationDir: dir);
-        // clean up the zip folder now that the data has been extracted
-        await file.delete();
+        if (!Directory.fromUri(Uri.parse('$appDirPath/$nameWithoutExt'))
+            .existsSync()) {
+          final dir =
+              Directory.fromUri(Uri.parse('$appDirPath/$nameWithoutExt'));
+          await ZipFile.extractToDirectory(zipFile: file, destinationDir: dir);
+          // clean up the zip folder now that the data has been extracted
+          await file.delete();
+        }
       }
     }
   }
